carms-data-platform-demo/src/config.py
import os
from dotenv import load_dotenv

load_dotenv()


class Settings:
    DATABASE_URL: str = os.getenv(
        "DATABASE_URL",
        "postgresql+psycopg2://carms:carms@localhost:5432/carms_db"
    )


settings = Settings()
